@@ -17,16 +17,17 @@
   ],
   "devDependencies": {
     "@types/expect.js": "^0.3.29",
-    "@types/mocha": "^9.1.1",
-    "@types/node": "^18.7.16",
-    "@types/prompts": "^2.0.14",
+    "@types/mocha": "^10.0.0",
+    "@types/node": "^18.11.9",
+    "@types/prompts": "^2.4.1",
+    "@vitejs/plugin-react": "^2.2.0",
     "buffer": "^6.0.3",
     "cross-env": "^7.0.3",
     "expect.js": "^0.3.1",
-    "mocha": "^10.0.0",
-    "nodecg-bundler": "^1.3.1",
+    "mocha": "^10.1.0",
+    "nodecg-bundler": "^1.4.0-next",
     "nodecg-types": "^1.9.0",
-    "npm-check-updates": "^16.1.2",
+    "npm-check-updates": "^16.4.1",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.7.1",
     "prompts": "^2.4.2",
@@ -35,20 +36,23 @@
     "ts-dev": "^2.1.16",
     "ts-mocha": "^10.0.0",
     "ts-node": "^10.9.1",
-    "tslib": "^2.4.0",
-    "turbo": "^1.4.6",
-    "typescript": "^4.8.3",
-    "vite": "^3.1.0"
+    "tslib": "^2.4.1",
+    "turbo": "^1.6.3",
+    "typescript": "^4.9.3",
+    "vite": "^3.2.4"
   },
   "dependencies": {
-    "@emotion/react": "^11.10.4",
+    "@emotion/react": "^11.10.5",
     "@iconify/react": "^4.0.0",
-    "@mantine/core": "^5.3.0",
-    "@mantine/form": "^5.3.0",
-    "@mantine/hooks": "^5.3.0",
+    "@mantine/core": "^5.8.2",
+    "@mantine/form": "^5.8.2",
+    "@mantine/hooks": "^5.8.2",
+    "@mantine/modals": "^5.8.2",
+    "cheerio": "^1.0.0-rc.12",
     "faker": "^6.6.6",
     "fakeyou.js": "^1.1.1",
-    "postcss": "^8.4.16",
-    "tailwindcss": "^3.1.8"
+    "postcss": "^8.4.19",
+    "tailwindcss": "^3.2.4",
+    "use-nodecg": "^0.4.1-1"
   }
 }
