@@ -7,15 +7,15 @@
   },
   "description": "Template bundle for bundling with parcel.",
   "devDependencies": {
-    "@types/node": "^18.7.16",
-    "autoprefixer": "^10.4.8",
+    "@types/node": "^18.11.9",
+    "autoprefixer": "^10.4.13",
     "npm-run-all": "^4.1.5",
-    "postcss": "^8.4.16",
+    "postcss": "^8.4.19",
     "postcss-import": "^15.0.0",
-    "postcss-nesting": "^10.1.10",
+    "postcss-nesting": "^10.2.0",
     "postcss-url": "^10.1.3",
-    "tailwindcss": "^3.1.8",
-    "typescript": "^4.8.3"
+    "tailwindcss": "^3.2.4",
+    "typescript": "^4.9.3"
   },
   "files": [
     "dashboard",
