@@ -54,23 +54,24 @@
     "clean": "tsc -b --clean"
   },
   "dependencies": {
-    "@types/node": "^18.7.16",
-    "gsap": "^3.11.1",
+    "@types/node": "^18.11.9",
+    "gsap": "^3.11.3",
     "nodecg-types": "^1.9.0",
-    "obs-websocket-js": "^5.0.1",
-    "postcss": "^8.4.16",
-    "typescript": "^4.8.3"
+    "obs-websocket-js": "^5.0.2",
+    "postcss": "^8.4.19",
+    "typescript": "^4.9.3"
   },
   "repository": "https://github.com/devJimmyboy/stream-overlay.git",
   "license": "Apache-2.0",
   "devDependencies": {
-    "autoprefixer": "^10.4.8",
-    "daisyui": "^2.25.0",
+    "@vitejs/plugin-react": "^2.2.0",
+    "autoprefixer": "^10.4.13",
+    "daisyui": "^2.41.0",
     "nodecg": "^1.9.0",
-    "postcss": "^8.4.16",
+    "postcss": "^8.4.19",
     "postcss-import": "^15.0.0",
-    "postcss-nesting": "^10.1.10",
+    "postcss-nesting": "^10.2.0",
     "postcss-url": "^10.1.3",
-    "tailwindcss": "^3.1.8"
+    "tailwindcss": "^3.2.4"
   }
 }
