@@ -1,34 +1,35 @@
 {
   "browserslist": "chrome >= 95",
   "dependencies": {
-    "@eastdesire/jscolor": "^2.5.0",
+    "@eastdesire/jscolor": "^2.5.1",
     "@iconify/iconify": "^3.0.0",
     "animate.css": "^4.1.1",
-    "axios": "^0.27.2",
+    "axios": "^1.1.3",
     "bad-words": "^3.0.4",
-    "daisyui": "^2.25.0",
-    "gsap": "^3.11.1",
+    "daisyui": "^2.41.0",
+    "gsap": "^3.11.3",
     "jquery": "^3.6.1",
     "lodash": "^4.17.21",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-select": "^5.4.0",
+    "react-select": "^5.6.1",
+    "uberduck-api": "^1.2.2",
     "use-nodecg": "^0.4.1-1"
   },
   "devDependencies": {
-    "@types/express": "^4.17.13",
-    "@types/lodash": "^4.14.184",
-    "@types/node": "^18.7.16",
-    "autoprefixer": "^10.4.8",
+    "@types/express": "^4.17.14",
+    "@types/lodash": "^4.14.189",
+    "@types/node": "^18.11.9",
+    "autoprefixer": "^10.4.13",
     "nodecg-io-streamelements": "^0.2.0",
     "nodecg-types": "^1.9.0",
     "npm-run-all": "^4.1.5",
-    "postcss": "^8.4.16",
+    "postcss": "^8.4.19",
     "postcss-import": "^15.0.0",
-    "postcss-nesting": "^10.1.10",
+    "postcss-nesting": "^10.2.0",
     "postcss-url": "^10.1.3",
-    "tailwindcss": "^3.1.8",
-    "typescript": "^4.8.3"
+    "tailwindcss": "^3.2.4",
+    "typescript": "^4.9.3"
   },
   "files": [
     "dashboard",
@@ -76,6 +77,14 @@
         "name": "config-panel",
         "title": "Bad Words Config",
         "width": 2,
+        "workspace": "config"
+      },
+      {
+        "file": "ttsTest.html",
+        "headerColor": "#525F78",
+        "name": "tts-test",
+        "title": "TTS Testing",
+        "width": 3,
         "workspace": "config"
       }
     ],
