@@ -1,18 +1,19 @@
 import { MantineProvider } from "@mantine/core";
 import React from "react";
-import ReactDOM from "react-dom";
+import ReactDOM from "react-dom/client";
 import App from "./components/App";
 import "./panel.css";
 
 // 2. Add your color mode confi
 
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById("root")!);
+
+root.render(
   <MantineProvider
     withGlobalStyles
     withNormalizeCSS
     theme={{ colorScheme: "dark" }}
   >
     <App />
-  </MantineProvider>,
-  document.getElementById("root")
+  </MantineProvider>
 );
