@@ -31,12 +31,14 @@ export default function (
 
   function onTip(data) {
     alertHandler.sendAlert(
-      data.type,
-      `(${data.name}) just SENT ME MONEY!!!!! - (${data.userCurrency.symbol}${data.amount} Pog Dollars)`,
+      "tip",
+      `(${data.name}) just SENT ME MONEY!!!!! - (${
+        data.currency === "USD" ? "$" : ""
+      }${data.amount}) (Pog) (Dollars)`,
       data.message
     );
   }
-  const jwt = `eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjoiNWZkMDQ1YTE2YmYwODE0YTk1NTM2YWNiIiwicm9sZSI6Im93bmVyIiwiY2hhbm5lbCI6IjVmZDA0NWExNmJmMDgxMjcxNjUzNmFjYyIsInByb3ZpZGVyIjoidHdpdGNoIiwiYXV0aFRva2VuIjoiNkdlTkUzc2lqSWlsYlNDdVBtdVlqRmFKbXV6OG91NTJKcHc0U2pzYU9NeDd3RW9NIiwiaWF0IjoxNjU1ODAxNjYxLCJpc3MiOiJTdHJlYW1FbGVtZW50cyJ9.m3sbigWyXnI5FkNl_JbNR56PUmbs6L03Gh0e4EN0_94`;
+  const jwt = process.env["STREAMELEMENTS_JWT"];
 
   socket.on("connect", onConnect);
   // Socket got disconnected
@@ -45,33 +47,43 @@ export default function (
   socket.on("authenticated", onAuthenticated);
   socket.on("unauthorized", console.error);
   socket.on("event:test", (data) => {
-    console.log(data);
+    nodecg.log.info(data);
+    onEvent(data, true);
     // Structure as on https://github.com/StreamElements/widgets/blob/master/CustomCode.md#on-event
   });
   socket.on("event", (data) => {
-    console.log(data);
+    nodecg.log.info(data);
+    onEvent(data);
     // Structure as on https://github.com/StreamElements/widgets/blob/master/CustomCode.md#on-event
   });
+  socket.on("event:update", (data) => {
+    nodecg.log.info(data);
+    // Structure as on https://github.com/StreamElements/widgets/blob/master/CustomCode.md#on-session-update
+  });
+  socket.on("event:reset", (data) => {
+    nodecg.log.info(data);
+    // Structure as on https://github.com/StreamElements/widgets/blob/master/CustomCode.md#on-session-update
+  });
 
-  function onEvent(event: any) {
-    const type = event.detail.listener;
+  function onEvent(event: any, test = false) {
+    const type = event.listener;
     if (type === "tip-latest") {
-      onTip(event.detail.event);
+      onTip(event.event);
     }
   }
   function onConnect() {
-    console.log("Successfully connected to the websocket");
+    nodecg.log.info("Successfully connected to the websocket");
     socket.emit("authenticate", { method: "jwt", token: jwt });
   }
 
   function onDisconnect() {
-    console.log("Disconnected from websocket");
+    nodecg.log.info("Disconnected from websocket");
     // Reconnect
     socket.connect();
   }
 
   function onAuthenticated(data) {
     const { channelId } = data;
-    console.log(`Successfully connected to channel ${channelId}`);
+    nodecg.log.info(`Successfully connected to channel ${channelId}`);
   }
 }
